refactor(Stepper): move to @xds/lab — API needs rework
The Stepper component needs significant design rework before it belongs
in core:

- Navigational semantics need to reference patterns like tabs/aria-current
- Status set should be limited to the global semantic set (success/error/warning)
- Icons should be a composable slot on each step, not baked-in via indicator modes
- Progress bar colors should use neutral/accent semantics, not icon-primary
- No React introspection (Children.toArray) allowed

Moving to lab while the API is refined. The component should do less,
not more — simpler primitives that compose well.
