+++ packages/core/src/index.ts
@@ -97,7 +97,6 @@ export * from './SideNav';
 export * from './MobileNav';
 export * from './Pagination';
 export * from './ProgressBar';
-export * from './Stepper';
 
 // Layout utilities and components (includes XDSHStack, XDSVStack)
 export * from './Layout';
