+++ packages/lab/src/index.ts
@@ -237,3 +237,4 @@ export {
   type StreamGLOptions,
   type StreamGLHandle,
 } from './ChartV2';
+export * from './Stepper';
